remove duplicated files and split types file into many
