+++ a2asrv/handler.go
@@ -10,19 +10,19 @@ import (
 type RequestHandler interface {
 	OnHandleGetTask(ctx context.Context, query a2a.TaskQueryParams) (a2a.Task, error)
 
-	OnHandleCancelTask(ctx context.Context, id a2a.TaskIDParams) (a2a.CancelTaskResponse, error)
+	OnHandleCancelTask(ctx context.Context, id a2a.TaskIDParams) (a2a.Task, error)
 
-	OnHandleSendMessage(ctx context.Context, message a2a.MessageSendParams) (a2a.SendMessageResponse, error)
+	OnHandleSendMessage(ctx context.Context, message a2a.MessageSendParams) (a2a.SendMessageResult, error)
 
 	OnHandleResubscribeToTask(ctx context.Context, id a2a.TaskIDParams) iter.Seq2[a2a.Event, error]
 
 	OnHandleSendMessageStream(ctx context.Context, message a2a.MessageSendParams) iter.Seq2[a2a.Event, error]
 
-	OnHandleGetTaskPushConfig(ctx context.Context, params a2a.GetTaskPushNotificationConfigParams) (a2a.GetTaskPushNotificationConfigResponse, error)
+	OnHandleGetTaskPushConfig(ctx context.Context, params a2a.GetTaskPushConfigParams) (a2a.TaskPushConfig, error)
 
-	OnHandleListTaskPushConfig(ctx context.Context, params a2a.ListTaskPushNotificationConfigParams) (a2a.ListTaskPushNotificationConfigResponse, error)
+	OnHandleListTaskPushConfig(ctx context.Context, params a2a.ListTaskPushConfigParams) ([]a2a.TaskPushConfig, error)
 
-	OnHandleSetTaskPushConfig(ctx context.Context, params a2a.SetTaskPushNotificationConfigParams) (a2a.SetTaskPushNotificationConfigResponse, error)
+	OnHandleSetTaskPushConfig(ctx context.Context, params a2a.SetTaskPushConfigParams) (a2a.SetTaskPushConfigResult, error)
 
-	OnHandleDeleteTaskPushConfig(ctx context.Context, params a2a.DeleteTaskPushNotificationConfigParams) (a2a.DeleteTaskPushNotificationConfigResponse, error)
+	OnHandleDeleteTaskPushConfig(ctx context.Context, params a2a.DeleteTaskPushConfigParams) error
 }
